pyproject.toml
[tool.poetry]
name = "medical-summary-assistant"
version = "0.1.0"
description = "FastAPI backend for the Medical Summary Assistant"
authors = ["Codex <codex@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
pyyaml = "^6.0"
spacy = "^3.7.2"
scispacy = "^0.5.4"
langgraph = "^0.1.3"
langchain-core = "^0.1.28"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[tool.poetry.virtualenvs]
in-project = true
path = ".venv"
create = true

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
